+++ src/main/java/org/bf/reportservice/application/command/ReportModifyService.java
@@ -1,0 +1,12 @@
+package org.bf.reportservice.application.command;
+
+import org.bf.reportservice.presentation.dto.ReportDeleteResponse;
+import org.bf.reportservice.presentation.dto.ReportResponse;
+import org.bf.reportservice.presentation.dto.ReportUpdateRequest;
+
+import java.util.UUID;
+
+public interface ReportModifyService {
+    ReportResponse update(UUID reportId, UUID userId, ReportUpdateRequest request);
+    ReportDeleteResponse delete(UUID reportId, UUID requesterUserId);
+}
