+++ src/main/java/org/bf/reportservice/application/event/dto/ReportImageLocation.java
@@ -1,0 +1,6 @@
+package org.bf.reportservice.application.event.dto;
+
+public record ReportImageLocation(
+        Double latitude,
+        Double longitude
+) {}
